# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TERDEPAN API Documentation",
        "description": "Dokumentasi API Sistem TERDEPAN (e-Kinerja BAPPEDA)",
        "contact": {
            "email": "dev@terdepan.local"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API v1"
        }
    ],
    "paths": {
...
        "/users": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "List semua pegawai",
                "operationId": "119035f2212e07ed2b4f988c4cdd59e2",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Tambah pegawai",
                "description": "Menambah data pegawai baru (users + user_profiles).",
                "operationId": "c16daeb14c1bd6b0d688a28578a43755",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nomor_identitas",
                                    "tipe_identitas",
                                    "nama",
                                    "jabatan_id",
                                    "bidang_id",
                                    "jenis_kelamin",
                                    "status_kepegawaian",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "nomor_identitas": {
                                        "type": "string",
                                        "example": "197812312005011001"
                                    },
                                    "tipe_identitas": {
                                        "type": "string",
                                        "example": "NIP"
                                    },
                                    "nama": {
                                        "type": "string",
                                        "example": "Ahmad Pegawai"
                                    },
                                    "jabatan_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "bidang_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "jenis_kelamin": {
                                        "type": "string",
                                        "enum": [
                                            "L",
                                            "P"
                                        ],
                                        "example": "L"
                                    },
                                    "status_kepegawaian": {
                                        "type": "string",
                                        "example": "Kontrak"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "pegawai@email.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Pegawai berhasil ditambah"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "nama": {
                                                    "type": "string",
                                                    "example": "Ahmad Pegawai"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "pegawai@email.com"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}